@@ -23,28 +23,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.6",
-    "@angular/cdk": "5.2.2",
-    "@angular/common": "5.2.6",
-    "@angular/compiler": "5.2.6",
-    "@angular/core": "5.2.6",
-    "@angular/flex-layout": "v2.0.0-beta.12",
-    "@angular/forms": "5.2.6",
-    "@angular/http": "5.2.6",
-    "@angular/material": "5.2.2",
-    "@angular/material-moment-adapter": "5.2.2",
-    "@angular/platform-browser": "5.2.6",
-    "@angular/platform-browser-dynamic": "5.2.6",
-    "@angular/platform-server": "5.2.6",
-    "@angular/router": "5.2.6",
-    "@ngrx/effects": "5.1.0",
-    "@ngrx/entity": "5.1.0",
-    "@ngrx/router-store": "5.0.1",
-    "@ngrx/store": "5.1.0",
-    "@ngrx/store-devtools": "5.1.0",
-    "@nguniversal/express-engine": "5.0.0-beta.5",
-    "@nguniversal/module-map-ngfactory-loader": "5.0.0-beta.5",
-    "angular-calendar": "0.23.6",
+    "@angular/animations": "6.1.4",
+    "@angular/cdk": "6.4.6",
+    "@angular/common": "6.1.4",
+    "@angular/compiler": "6.1.4",
+    "@angular/core": "6.1.4",
+    "@angular/flex-layout": "6.0.0-beta.16",
+    "@angular/forms": "6.1.4",
+    "@angular/http": "6.1.4",
+    "@angular/material": "6.4.6",
+    "@angular/material-moment-adapter": "6.4.6",
+    "@angular/platform-browser": "6.1.4",
+    "@angular/platform-browser-dynamic": "6.1.4",
+    "@angular/platform-server": "6.1.4",
+    "@angular/router": "6.1.4",
+    "@ngrx/effects": "6.0.1",
+    "@ngrx/entity": "6.0.1",
+    "@ngrx/router-store": "6.0.1",
+    "@ngrx/store": "6.0.1",
+    "@ngrx/store-devtools": "6.0.1",
+    "@nguniversal/common": "v6.1.0",
+    "@nguniversal/express-engine": "v6.1.0",
+    "@nguniversal/module-map-ngfactory-loader": "v6.1.0",
+    "angular-calendar": "0.25.2",
     "core-js": "2.5.3",
     "crypto-js": "3.1.9-1",
     "date-fns": "1.29.0",
@@ -68,23 +72,25 @@
     "@types/jasminewd2": "~2.0.3",
     "@types/lodash": "4.14.104",
     "@types/node": "~9.4.6",
-    "codelyzer": "~4.1.0",
+    "codelyzer": "~4.2.1",
     "concurrently": "3.5.1",
+    "http-server": "^0.11.1",
     "jasmine-core": "~2.99.1",
-    "jasmine-marbles": "0.2.0",
+    "jasmine-marbles": "0.3.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.1",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
     "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "ngrx-store-freeze": "0.2.1",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ngrx-store-freeze": "0.2.4",
     "protractor": "~5.3.0",
-    "rxjs-tslint-rules": "3.14.0",
+    "rxjs-tslint-rules": "~4.7.2",
     "ts-node": "~5.0.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
+    "typescript": "2.9.2",
+    "udk": "^0.3.17",
+    "webpack-cli": "^3.1.0",
     "webpack-node-externals": "1.6.0"
   },
   "description": "The Task Manager System",
