@@ -1,6 +1,6 @@
-import {Injectable} from '@angular/core';
-import {BehaviorSubject} from 'rxjs/BehaviorSubject';
-import {Observable} from 'rxjs/Observable';
+import { Injectable } from '@angular/core';
+import { BehaviorSubject } from 'rxjs';
+import { Observable } from 'rxjs';
 
 export interface DragData {
   tag: string; // 用于标识该拖拽对象，在具有多个可拖拽的层级中标识该层级，需要用户自己维护唯一性
@@ -9,7 +9,6 @@ export interface DragData {
 
 @Injectable()
 export class DragDropService {
-
   private _dragData = new BehaviorSubject<DragData | null>(null);
 
   setDragData(data: DragData) {
