@@ -1,10 +1,10 @@
-import {Injectable} from '@angular/core';
-import {Actions, Effect, ofType} from '@ngrx/effects';
-import {Action} from '@ngrx/store';
-import {Observable} from 'rxjs/Observable';
-import {of} from 'rxjs/observable/of';
+import { Injectable } from '@angular/core';
+import { Actions, Effect, ofType } from '@ngrx/effects';
+import { Action } from '@ngrx/store';
+import { Observable } from 'rxjs';
+import { of } from 'rxjs';
 import { switchMap, map, catchError } from 'rxjs/operators';
-import {QuoteService} from '../services';
+import { QuoteService } from '../services';
 import * as actions from '../actions/quote.action';
 
 @Injectable()
@@ -13,16 +13,15 @@ export class QuoteEffects {
    *
    */
   @Effect()
-  quote$: Observable<Action> = this.actions$
-    .pipe(
-      ofType<actions.QuoteAction>(actions.QUOTE),
-      switchMap(() => this.quoteService.getQuote()
-        .pipe(
-          map(quote => new actions.QuoteSuccessAction(quote)),
-          catchError(err => of(new actions.QuoteFailAction(JSON.stringify(err))))
-        )
+  quote$: Observable<Action> = this.actions$.pipe(
+    ofType<actions.QuoteAction>(actions.QUOTE),
+    switchMap(() =>
+      this.quoteService.getQuote().pipe(
+        map(quote => new actions.QuoteSuccessAction(quote)),
+        catchError(err => of(new actions.QuoteFailAction(JSON.stringify(err))))
       )
-    );
+    )
+  );
 
   /**
    *
