@@ -1,16 +1,16 @@
-import {Component, HostBinding, ChangeDetectionStrategy} from '@angular/core';
-import {MatDialog} from '@angular/material';
-import {Store, select} from '@ngrx/store';
-import {Observable} from 'rxjs/Observable';
+import { Component, HostBinding, ChangeDetectionStrategy } from '@angular/core';
+import { MatDialog } from '@angular/material';
+import { Store, select } from '@ngrx/store';
+import { Observable } from 'rxjs';
 import * as fromRoot from '../../reducers';
 import * as actions from '../../actions/project.action';
-import {NewProjectComponent} from '../components/new-project';
-import {InviteComponent} from '../components/invite';
-import {ConfirmDialogComponent} from '../../shared';
-import {defaultRouteAnim, listAnimation} from '../../anim';
+import { NewProjectComponent } from '../components/new-project';
+import { InviteComponent } from '../components/invite';
+import { ConfirmDialogComponent } from '../../shared';
+import { defaultRouteAnim, listAnimation } from '../../anim';
 import { Project, User } from '../../domain';
 import { map, take, switchMap, reduce, filter } from 'rxjs/operators';
-import { range } from 'rxjs/observable/range';
+import { range } from 'rxjs';
 
 @Component({
   selector: 'app-project-list',
@@ -33,28 +33,32 @@ import { range } from 'rxjs/observable/range';
       <mat-icon>add</mat-icon>
     </button>
   `,
-  styles: [`
-    .card {
-      margin: 10px;
-    }
-    .fab-button {
-      position: fixed;
-      right: 32px;
-      bottom: 96px;
-      z-index: 998;
-    }
-  `],
+  styles: [
+    `
+      .card {
+        margin: 10px;
+      }
+      .fab-button {
+        position: fixed;
+        right: 32px;
+        bottom: 96px;
+        z-index: 998;
+      }
+    `
+  ],
   animations: [defaultRouteAnim, listAnimation],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ProjectListComponent {
-
-  @HostBinding('@routeAnim') state: string;
+  @HostBinding('@routeAnim')
+  state: string;
   projects$: Observable<Project[]>;
   listAnim$: Observable<number>;
 
-  constructor(private store$: Store<fromRoot.State>,
-              private dialog: MatDialog) {
+  constructor(
+    private store$: Store<fromRoot.State>,
+    private dialog: MatDialog
+  ) {
     this.store$.dispatch(new actions.LoadProjectsAction());
     this.projects$ = this.store$.pipe(select(fromRoot.getProjects));
     this.listAnim$ = this.projects$.pipe(map(p => p.length));
@@ -67,30 +71,40 @@ export class ProjectListComponent {
   openNewProjectDialog() {
     const img = `/assets/img/covers/${Math.floor(Math.random() * 40)}_tn.jpg`;
     const thumbnails$ = this.getThumbnailsObs();
-    const dialogRef = this.dialog.open(NewProjectComponent, {data: { thumbnails: thumbnails$, img: img}});
-    dialogRef.afterClosed()
-      .pipe(
-        take(1)
-      )
+    const dialogRef = this.dialog.open(NewProjectComponent, {
+      data: { thumbnails: thumbnails$, img: img }
+    });
+    dialogRef
+      .afterClosed()
+      .pipe(take(1))
       .subscribe(val => {
         if (val) {
           const converImg = this.buildImgSrc(val.coverImg);
-          this.store$.dispatch(new actions.AddProjectAction({...val, coverImg: converImg}));
+          this.store$.dispatch(
+            new actions.AddProjectAction({ ...val, coverImg: converImg })
+          );
         }
       });
   }
 
   openUpdateDialog(project: Project) {
     const thumbnails$ = this.getThumbnailsObs();
-    const dialogRef = this.dialog.open(NewProjectComponent, {data: { project: project, thumbnails: thumbnails$}});
-    dialogRef.afterClosed()
-      .pipe(
-        take(1)
-      )
+    const dialogRef = this.dialog.open(NewProjectComponent, {
+      data: { project: project, thumbnails: thumbnails$ }
+    });
+    dialogRef
+      .afterClosed()
+      .pipe(take(1))
       .subscribe(val => {
         if (val) {
           const converImg = this.buildImgSrc(val.coverImg);
-          this.store$.dispatch(new actions.UpdateProjectAction({...val, id: project.id, coverImg: converImg}));
+          this.store$.dispatch(
+            new actions.UpdateProjectAction({
+              ...val,
+              id: project.id,
+              coverImg: converImg
+            })
+          );
         }
       });
   }
@@ -100,16 +114,23 @@ export class ProjectListComponent {
       .pipe(
         select(fromRoot.getProjectMembers(<string>project.id)),
         take(1),
-        map(members => this.dialog.open(InviteComponent, {data: { members: members}})),
-        switchMap(dialogRef => dialogRef.afterClosed()
-          .pipe(
+        map(members =>
+          this.dialog.open(InviteComponent, { data: { members: members } })
+        ),
+        switchMap(dialogRef =>
+          dialogRef.afterClosed().pipe(
             take(1),
             filter(n => n)
           )
         )
       )
       .subscribe(val => {
-        this.store$.dispatch(new actions.InviteMembersAction({projectId: <string>project.id, members: <User[]>val}));
+        this.store$.dispatch(
+          new actions.InviteMembersAction({
+            projectId: <string>project.id,
+            members: <User[]>val
+          })
+        );
       });
   }
 
@@ -119,13 +140,14 @@ export class ProjectListComponent {
       content: '确认要删除该项目？',
       confirmAction: '确认删除'
     };
-    const dialogRef = this.dialog.open(ConfirmDialogComponent, {data: {dialog: confirm}});
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: { dialog: confirm }
+    });
 
     // 使用 take(1) 来自动销毁订阅，因为 take(1) 意味着接收到 1 个数据后就完成了
-    dialogRef.afterClosed()
-      .pipe(
-        take(1)
-      )
+    dialogRef
+      .afterClosed()
+      .pipe(take(1))
       .subscribe(val => {
         if (val) {
           this.store$.dispatch(new actions.DeleteProjectAction(project));
@@ -134,13 +156,12 @@ export class ProjectListComponent {
   }
 
   private getThumbnailsObs(): Observable<string[]> {
-    return range(0, 40)
-      .pipe(
-        map(i => `/assets/img/covers/${i}_tn.jpg`),
-        reduce((r: string[], x: string) => {
-          return [...r, x];
-        }, [])
-      );
+    return range(0, 40).pipe(
+      map(i => `/assets/img/covers/${i}_tn.jpg`),
+      reduce((r: string[], x: string) => {
+        return [...r, x];
+      }, [])
+    );
   }
 
   private buildImgSrc(img: string): string {
