@@ -1,6 +1,6 @@
 import { Inject, Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { Task } from '../domain';
 import { CalendarEvent } from 'angular-calendar';
 import { endOfDay, startOfDay } from 'date-fns';
