@@ -1,19 +1,19 @@
 import { Inject, Injectable } from '@angular/core';
 import { HttpHeaders, HttpClient, HttpParams } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import { from } from 'rxjs/observable/from';
+import { Observable } from 'rxjs';
+import { from } from 'rxjs';
 import { mergeMap, reduce, mapTo } from 'rxjs/operators';
 import { Task, User, TaskList } from '../domain';
 
 @Injectable()
 export class TaskService {
   private readonly domain = 'tasks';
-  private headers = new HttpHeaders()
-    .set('Content-Type', 'application/json');
+  private headers = new HttpHeaders().set('Content-Type', 'application/json');
 
-  constructor(@Inject('BASE_CONFIG') private config: { uri: string },
-    private http: HttpClient) {
-  }
+  constructor(
+    @Inject('BASE_CONFIG') private config: { uri: string },
+    private http: HttpClient
+  ) {}
 
   add(task: Task): Observable<Task> {
     const uri = `${this.config.uri}/${this.domain}`;
@@ -30,9 +30,9 @@ export class TaskService {
       createDate: task.createDate
     };
     // const addTaskRef$ = this.addTaskRef()
-    return this.http
-      .post<Task>(uri, JSON.stringify(toAdd), { headers: this.headers });
-
+    return this.http.post<Task>(uri, JSON.stringify(toAdd), {
+      headers: this.headers
+    });
   }
 
   update(task: Task): Observable<Task> {
@@ -46,70 +46,78 @@ export class TaskService {
       priority: task.priority,
       remark: task.remark
     };
-    return this.http
-      .patch<Task>(uri, JSON.stringify(toUpdate), { headers: this.headers });
+    return this.http.patch<Task>(uri, JSON.stringify(toUpdate), {
+      headers: this.headers
+    });
   }
 
   del(task: Task): Observable<Task> {
     const uri = `${this.config.uri}/${this.domain}/${task.id}`;
-    return this.http.delete(uri)
-      .pipe(
-        mapTo(task)
-      );
+    return this.http.delete(uri).pipe(mapTo(task));
   }
 
   // GET /tasklist
   get(taskListId: string): Observable<Task[]> {
     const uri = `${this.config.uri}/${this.domain}`;
-    const params = new HttpParams()
-      .set('taskListId', taskListId);
-    return this.http
-      .get<Task[]>(uri, { params });
+    const params = new HttpParams().set('taskListId', taskListId);
+    return this.http.get<Task[]>(uri, { params });
   }
 
   getByLists(lists: TaskList[]): Observable<Task[]> {
-    return from(lists)
-      .pipe(
-        mergeMap((list: TaskList) => this.get(<string>list.id)),
-        reduce((tasks: Task[], t: Task[]) => [...tasks, ...t], [])
-      );
+    return from(lists).pipe(
+      mergeMap((list: TaskList) => this.get(<string>list.id)),
+      reduce((tasks: Task[], t: Task[]) => [...tasks, ...t], [])
+    );
   }
 
   moveAll(srcListId: string, targetListId: string): Observable<Task[]> {
-    return this.get(srcListId)
-      .pipe(
-        mergeMap((tasks: Task[]) => from(tasks)),
-        mergeMap((task: Task) => this.move(<string>task.id, targetListId)),
-        reduce((arrTasks: Task[], t: Task) => {
-          return [...arrTasks, t];
-        }, [])
-      );
+    return this.get(srcListId).pipe(
+      mergeMap((tasks: Task[]) => from(tasks)),
+      mergeMap((task: Task) => this.move(<string>task.id, targetListId)),
+      reduce((arrTasks: Task[], t: Task) => {
+        return [...arrTasks, t];
+      }, [])
+    );
   }
 
   move(taskId: string, taskListId: string): Observable<Task> {
     const uri = `${this.config.uri}/${this.domain}/${taskId}`;
-    return this.http
-      .patch<Task>(uri, JSON.stringify({ taskListId: taskListId }), { headers: this.headers });
+    return this.http.patch<Task>(
+      uri,
+      JSON.stringify({ taskListId: taskListId }),
+      { headers: this.headers }
+    );
   }
 
   complete(task: Task): Observable<Task> {
     const uri = `${this.config.uri}/${this.domain}/${task.id}`;
-    return this.http
-      .patch<Task>(uri, JSON.stringify({ completed: !task.completed }), { headers: this.headers });
+    return this.http.patch<Task>(
+      uri,
+      JSON.stringify({ completed: !task.completed }),
+      { headers: this.headers }
+    );
   }
 
   addTaskRef(user: User, taskId: string): Observable<User> {
     const uri = `${this.config.uri}/users/${user.id}`;
-    const taskIds = (user.taskIds) ? user.taskIds : [];
-    return this.http
-      .patch<User>(uri, JSON.stringify({ taskIds: [...taskIds, taskId] }), { headers: this.headers });
+    const taskIds = user.taskIds ? user.taskIds : [];
+    return this.http.patch<User>(
+      uri,
+      JSON.stringify({ taskIds: [...taskIds, taskId] }),
+      { headers: this.headers }
+    );
   }
 
   removeTaskRef(user: User, taskId: string): Observable<User> {
     const uri = `${this.config.uri}/users/${user.id}`;
-    const taskIds = (user.taskIds) ? user.taskIds : [];
+    const taskIds = user.taskIds ? user.taskIds : [];
     const index = taskIds.indexOf(taskId);
-    return this.http
-      .patch<User>(uri, JSON.stringify({ taskIds: [...taskIds.slice(0, index), taskIds.slice(index + 1)] }), { headers: this.headers });
+    return this.http.patch<User>(
+      uri,
+      JSON.stringify({
+        taskIds: [...taskIds.slice(0, index), taskIds.slice(index + 1)]
+      }),
+      { headers: this.headers }
+    );
   }
 }
