@@ -1,6 +1,5 @@
-import {Injectable} from '@angular/core';
-import {Subject} from 'rxjs/Subject';
-
+import { Injectable } from '@angular/core';
+import { Subject } from 'rxjs';
 
 /** Datepicker data that requires internationalization. */
 @Injectable()
