@@ -1,12 +1,23 @@
-import {ChangeDetectionStrategy, Component, forwardRef, OnInit, OnDestroy} from '@angular/core';
-import {ControlValueAccessor, FormControl, NG_VALIDATORS, NG_VALUE_ACCESSOR} from '@angular/forms';
-import {Subject} from 'rxjs/Subject';
-import {Observable} from 'rxjs/Observable';
-import {Subscription} from 'rxjs/Subscription';
-import {Identity, IdentityType} from '../../domain';
-import {isValidAddr, extractInfo} from '../../utils/identity.util';
-import {isValidDate} from '../../utils/date.util';
-import { combineLatest } from 'rxjs/observable/combineLatest';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  forwardRef,
+  OnInit,
+  OnDestroy
+} from '@angular/core';
+import {
+  ControlValueAccessor,
+  FormControl,
+  NG_VALIDATORS,
+  NG_VALUE_ACCESSOR
+} from '@angular/forms';
+import { Subject } from 'rxjs';
+import { Observable } from 'rxjs';
+import { Subscription } from 'rxjs';
+import { Identity, IdentityType } from '../../domain';
+import { isValidAddr, extractInfo } from '../../utils/identity.util';
+import { isValidDate } from '../../utils/date.util';
+import { combineLatest } from 'rxjs';
 
 @Component({
   selector: 'app-indentity-input',
@@ -27,46 +38,48 @@ import { combineLatest } from 'rxjs/observable/combineLatest';
       </mat-form-field>
     </div>
   `,
-  styles: [`
-    .id-input{
-      flex: 1;
-    }
-    :host{
-      display: flex;
-      flex-direction: row;
-      flex-wrap: nowrap;
-      align-items: baseline;
-    }
-  `],
+  styles: [
+    `
+      .id-input {
+        flex: 1;
+      }
+      :host {
+        display: flex;
+        flex-direction: row;
+        flex-wrap: nowrap;
+        align-items: baseline;
+      }
+    `
+  ],
   providers: [
     {
       provide: NG_VALUE_ACCESSOR,
       useExisting: forwardRef(() => IdentityInputComponent),
-      multi: true,
+      multi: true
     },
     {
       provide: NG_VALIDATORS,
       useExisting: forwardRef(() => IdentityInputComponent),
-      multi: true,
+      multi: true
     }
   ],
-  changeDetection: ChangeDetectionStrategy.OnPush,
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class IdentityInputComponent implements ControlValueAccessor, OnInit, OnDestroy {
-
-  identityTypes: {value: IdentityType, label: string}[] = [
-    {value: IdentityType.IdCard, label: '身份证'},
-    {value: IdentityType.Insurance, label: '医保'},
-    {value: IdentityType.Passport, label: '护照'},
-    {value: IdentityType.Military, label: '军官证'},
-    {value: IdentityType.Other, label: '其它'}
+export class IdentityInputComponent
+  implements ControlValueAccessor, OnInit, OnDestroy {
+  identityTypes: { value: IdentityType; label: string }[] = [
+    { value: IdentityType.IdCard, label: '身份证' },
+    { value: IdentityType.Insurance, label: '医保' },
+    { value: IdentityType.Passport, label: '护照' },
+    { value: IdentityType.Military, label: '军官证' },
+    { value: IdentityType.Other, label: '其它' }
   ];
-  identity: Identity = {identityType: null, identityNo: null};
+  identity: Identity = { identityType: null, identityNo: null };
   private _idType = new Subject<IdentityType>();
   private _idNo = new Subject<string>();
   private _sub: Subscription;
 
-  constructor() { }
+  constructor() {}
 
   private propagateChange = (_: any) => {};
 
@@ -105,11 +118,10 @@ export class IdentityInputComponent implements ControlValueAccessor, OnInit, OnD
   }
 
   // 这里没有使用，用于注册 touched 状态
-  public registerOnTouched() {
-  }
+  public registerOnTouched() {}
 
   // 验证表单，验证结果正确返回 null 否则返回一个验证结果对象
-  validate(c: FormControl): {[key: string]: any} | null {
+  validate(c: FormControl): { [key: string]: any } | null {
     if (!c.value) {
       return null;
     }
@@ -130,11 +142,11 @@ export class IdentityInputComponent implements ControlValueAccessor, OnInit, OnD
     }
   }
 
-  private validateIdNumber(c: FormControl): {[key: string]: any} | null {
+  private validateIdNumber(c: FormControl): { [key: string]: any } | null {
     const val = c.value.identityNo;
     if (val.length !== 18) {
       return {
-        idNotValid:  true
+        idNotValid: true
       };
     }
     const pattern = /^[1-9]\d{5}[1-9]\d{3}((0\d)|(1[0-2]))(([0|1|2]\d)|3[0-1])\d{3}[x0-9]$/;
@@ -145,30 +157,30 @@ export class IdentityInputComponent implements ControlValueAccessor, OnInit, OnD
         result = true;
       }
     }
-    return result ? null : {idNotValid:  true};
+    return result ? null : { idNotValid: true };
   }
 
-  private validatePassport(c: FormControl): {[key: string]: any} | null {
+  private validatePassport(c: FormControl): { [key: string]: any } | null {
     const value = c.value.identityNo;
     if (value.length !== 9) {
-      return {idNotValid: true};
+      return { idNotValid: true };
     }
     const pattern = /^[GgEe]\d{8}$/;
     let result = false;
     if (pattern.test(value)) {
       result = true;
     }
-    return result ? null : {idNotValid:  true};
+    return result ? null : { idNotValid: true };
   }
 
-  private validateMilitary(c: FormControl): {[key: string]: any} | null {
+  private validateMilitary(c: FormControl): { [key: string]: any } | null {
     const value = c.value.identityNo;
     const pattern = /[\u4e00-\u9fa5](字第)(\d{4,8})(号?)$/;
     let result = false;
     if (pattern.test(value)) {
       result = true;
     }
-    return result ? null : {idNotValid:  true};
+    return result ? null : { idNotValid: true };
   }
 
   onIdTypeChange(idType: IdentityType) {
